마이페이지 관련 수정
