+++ src/main/java/game/dao/GameDAO.java
@@ -12,6 +12,7 @@ import javax.sql.DataSource;
 
 import commons.DBConfig;
 import game.dto.GameDTO;
+import mypage.dto.GameScoreDTO;
 
 public class GameDAO {
 	public static GameDAO instance;
@@ -70,6 +71,38 @@ public class GameDAO {
 				return dto;
 			}
 		}
+	}
+	
+	
+	/**
+	 * 게임 플레이 정보 출력
+	 * @param id
+	 * @return
+	 * @throws Exception
+	 */
+	public List<GameScoreDTO> gameList(String id) throws Exception {
+		
+		String sql = "select * from game_score gs join game g on g.seq = gs.game_seq where member_id = ? order by seq";
+		
+		try (Connection con=DBConfig.getConnection(); PreparedStatement pstat = con.prepareStatement(sql)) {
+
+			pstat.setString(1, id);
+			
+			try(ResultSet rs = pstat.executeQuery()){
+				
+				List<GameScoreDTO> list = new ArrayList<>();
+				while (rs.next()) {
+					String title = rs.getString("title");
+					int score = rs.getInt("score");
+					String member_id = rs.getString("member_id");
+					int game_seq = rs.getInt("game_seq");
+					
+					list.add(new GameScoreDTO(title, score, member_id, game_seq ));
+				} 
+				return list;
+			}
+		}
+		
 	}
